Simplify validation code
