--- genienlp/validate.py
@@ -1,125 +1,0 @@
-#
-# Copyright (c) 2018, Salesforce, Inc.
-#                     The Board of Trustees of the Leland Stanford Junior University
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-# * Redistributions of source code must retain the above copyright notice, this
-#   list of conditions and the following disclaimer.
-#
-# * Redistributions in binary form must reproduce the above copyright notice,
-#   this list of conditions and the following disclaimer in the documentation
-#   and/or other materials provided with the distribution.
-#
-# * Neither the name of the copyright holder nor the names of its
-#   contributors may be used to endorse or promote products derived from
-#   this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-import logging
-import sys
-
-import torch
-
-from .metrics import calculate_and_reduce_metrics
-
-logger = logging.getLogger(__name__)
-
-
-def generate_with_model(
-    model,
-    data_iterator,
-    task,
-    args,
-    output_predictions_only=False,
-    output_confidence_features=False,
-    original_order=None,
-    confidence_estimators=None,
-    disable_progbar=True,
-    eval_dir=None,
-):
-    if args.e2e_dialogue_evaluation:
-        return model.validate_e2e_dialogues(
-            data_iterator,
-            task,
-            eval_dir=eval_dir,
-            output_predictions_only=output_predictions_only,
-            original_order=original_order,
-            disable_progbar=disable_progbar,
-        )
-    else:
-        return model.validate(
-            data_iterator,
-            task,
-            output_predictions_only=output_predictions_only,
-            output_confidence_features=output_confidence_features,
-            original_order=original_order,
-            confidence_estimators=confidence_estimators,
-            disable_progbar=disable_progbar,
-        )
-
-
-def print_results(results, num_print):
-    print()
-
-    values = list(results.values())
-    num_examples = len(values[0])
-
-    # examples are sorted by length
-    # to get good diversity, get half of examples from second quartile
-    start = int(num_examples / 4)
-    end = start + int(num_print / 2)
-    first_list = [val[start:end] for val in values]
-
-    # and the other half from fourth quartile
-    start = int(3 * num_examples / 4)
-    end = start + num_print - int(num_print / 2)
-    second_list = [val[start:end] for val in values]
-
-    # join examples
-    processed_values = [first + second for first, second in zip(first_list, second_list)]
-
-    for ex_idx in range(len(processed_values[0])):
-        for key_idx, key in enumerate(results.keys()):
-            value = processed_values[key_idx][ex_idx]
-            v = value[0] if isinstance(value, list) else value
-            key_width = max(len(key) for key in results)
-            print(f'{key:>{key_width}}: {repr(v)}')
-        print()
-    sys.stdout.flush()
-
-
-def validate(task, val_iter, model, args, num_print=10):
-    with torch.no_grad():
-        model.eval()
-        if isinstance(model, torch.nn.DataParallel):
-            # get rid of the DataParallel wrapper
-            model = model.module
-
-        generation_output = generate_with_model(model, val_iter, task, args)
-
-        # loss is already calculated
-        metrics_to_return = [metric for metric in task.metrics if metric != 'loss']
-
-        metrics = calculate_and_reduce_metrics(args, generation_output, metrics_to_return, model.tgt_lang)
-
-        results = {
-            'model prediction': generation_output.predictions,
-            'gold answer': generation_output.answers,
-            'context': generation_output.contexts,
-        }
-
-        print_results(results, num_print)
-
-        return generation_output, metrics
